requirements.txt
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.1
defusedxml==0.7.1
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
isodate==0.6.0
kavenegar==1.1.2
lxml==4.6.3
Markdown==3.3.4
msgpack==1.0.2
msgpack-python==0.5.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2021.1
redis==2.10.6
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==0.9.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
urllib3==1.26.4
zeep==4.0.0
zope.interface==5.3.0
